==> notebooks/project/00.1_MESF_Preprocessing.ipynb <==
import fs
import pandas as pd
import numpy as np

RAW_PROJECT_DIR = fs.open_fs("../../data/raw/project")
INTERIM_PROJECT_DIR = fs.open_fs("../../data/interim/project")
RAW_FILE_NAME = RAW_PROJECT_DIR.getsyspath("hmeq.csv")
INTERIM_FILE_NAME = INTERIM_PROJECT_DIR.getsyspath("use_to_divide.csv")

### Dataset Description

data = pd.read_csv(RAW_FILE_NAME)

data.shape

data.dtypes

X = data.drop(columns=['BAD'])
y = data['BAD']

numeric_features = X.select_dtypes(include=['int64', 'float64']).columns
categoric_features = X.select_dtypes(include=['object']).columns

data[numeric_features].describe().T

X[categoric_features].describe().T

### Nulls

null_percentage = np.round(data.isnull().sum() * 100 / len(data), 2)

missing_values_df = pd.DataFrame({'missing_values_%': null_percentage})
missing_values_df

### Duplicate Data

data.duplicated().sum()

### Imputing Nulls

data['JOB'] = data['JOB'].fillna('Other')

medians_job = data.groupby('JOB')[['MORTDUE', 'VALUE', 'YOJ', 'DEROG', 'DELINQ', 'CLAGE', 'NINQ', 'CLNO', 'DEBTINC']].median()
medians_job = medians_job.reset_index()
medians_job = medians_job.rename(columns={'MORTDUE': 'MED_MORTDUE', 'VALUE': 'MED_VALUE', 
'YOJ': 'MED_YOJ', 'DEROG': 'MED_DEROG', 'DELINQ': 'MED_DELINQ', 'CLAGE': 'MED_CLAGE', 
'NINQ': 'MED_NINQ', 'CLNO': 'MED_CLNO', 'DEBTINC': 'MED_DEBTINC'})
mode_reason = data.groupby('JOB')[['REASON']].agg(lambda x: x.mode().iloc[0]).reset_index()
mode_reason = mode_reason.rename(columns={'REASON': 'MED_MODREASON'})

mode_reason

medians_job.head()

data_to_get_okey = pd.merge(data, medians_job, on='JOB', how='left')
data_to_get_okey = data_to_get_okey.fillna({'MORTDUE': data_to_get_okey['MED_MORTDUE'], 
'VALUE': data_to_get_okey['MED_VALUE'], 
'YOJ': data_to_get_okey['MED_YOJ'], 
'DEROG': data_to_get_okey['MED_DEROG'], 
'DELINQ': data_to_get_okey['MED_DELINQ'], 
'CLAGE': data_to_get_okey['MED_CLAGE'],
'NINQ': data_to_get_okey['MED_NINQ'], 
'CLNO': data_to_get_okey['MED_CLNO'],
'DEBTINC': data_to_get_okey['MED_DEBTINC']}).drop(data_to_get_okey.filter(regex='^MED_').columns, axis=1)
data_okey =  pd.merge(data_to_get_okey, mode_reason, on='JOB', how='left')
data_okey = data_okey.fillna({'REASON': data_okey['MED_MODREASON']}).drop(['MED_MODREASON'], axis=1)

null_percentage = np.round(data_okey.isnull().sum() * 100 / len(data_okey), 2)

missing_values_df = pd.DataFrame({'missing_values_%': null_percentage})
missing_values_df

data_okey

### Detecting Outliers

def find_outliers_iqr(df, columns=None):
    # If no specific columns are provided, use only numeric columns
    if columns is None:
        columns = df.select_dtypes(include=[np.number]).columns
    
    # Create an empty list to store whether a row contains any outlier
    outlier_flags = []
    
    # Precompute IQR bounds for each numeric column
    iqr_bounds = {}
    for col in columns:
        Q1 = df[col].quantile(0.01)
        Q3 = df[col].quantile(0.99)
        IQR = Q3 - Q1
        lower_bound = Q1 - 1.5 * IQR
        upper_bound = Q3 + 1.5 * IQR
        iqr_bounds[col] = (lower_bound, upper_bound)
    
    # Iterate over each row in the DataFrame
    for index, row in df.iterrows():
        row_outlier = False  # Initialize flag for each row
        
        # Check each numeric column for an outlier using the precomputed bounds
        for col in columns:
            lower_bound, upper_bound = iqr_bounds[col]
            
            # Check if the current value in this row and column is an outlier
            if row[col] < lower_bound or row[col] > upper_bound:
                row_outlier = True
                break  # Stop checking other columns for this row once an outlier is found
        
        # Append 'Outlier' or 'Not Outlier' based on the row_outlier flag
        outlier_flags.append('Outlier' if row_outlier else 'Not Outlier')
    
    # Add a new column to the DataFrame
    df['outlier_flag'] = outlier_flags
    
    return df

outliers = find_outliers_iqr(data_okey)

outliers

outliers.groupby('outlier_flag').size()

### Save Clean Data

outliers.to_csv(INTERIM_FILE_NAME, index=False, encoding="utf-8")
