==> notebooks/project/02.0_MESF_Feature_Selection.ipynb <==
import fs
import pandas as pd
from sklearn.preprocessing import MinMaxScaler, LabelEncoder
from sklearn.feature_selection import VarianceThreshold
from varclushi import VarClusHi
from optbinning import BinningProcess

PROCESSED_PROJECT_DIR = fs.open_fs("../../data/processed/project")
TRAIN_FILE_NAME = PROCESSED_PROJECT_DIR.getsyspath("train.csv")

train_df = pd.read_csv(TRAIN_FILE_NAME)

### Checking Variance

categoric_features = ['BAD', 'REASON', 'JOB']
train_df[categoric_features] = train_df[categoric_features].apply(lambda col: col.astype('category'))

numeric_features = train_df.select_dtypes(include=['int', 'float']).columns.tolist()

scaler = MinMaxScaler(feature_range=(10, 100))
train_df[numeric_features] = scaler.fit_transform(train_df[numeric_features])

vt = VarianceThreshold(threshold=0.15)

vt.fit(train_df[numeric_features])

varianza_peq = [v for v,nu in zip(train_df[numeric_features].columns,vt.get_support()) if not nu]
len(varianza_peq)

### Applying BinningProcess

X = train_df.drop(columns=['BAD'])
y = train_df['BAD']

categoric_features = ['REASON', 'JOB']

label_encoders = {}

for column in categoric_features:
    label_encoders[column] = LabelEncoder()
    X[column] = label_encoders[column].fit_transform(X[column])

vc = VarClusHi(df = X)
vc.varclus()

rs = vc.rsquare
rs = rs.sort_values(by=['Cluster', 'RS_Ratio']).reset_index(drop = True)
rs['id'] = rs.groupby('Cluster').cumcount()+1

rs

variable_names = list(X.columns)

binning_process = BinningProcess(variable_names)
binning_process.fit(X, y)
binning_process.information(print_level=1)

df_bins_num  = binning_process.summary().sort_values(by = 'iv', ascending = False)
df_bins_num  = df_bins_num[df_bins_num['dtype'] == 'numerical']
df_bins_num.rename(columns = {'name': 'Variable'}, inplace = True)
df_bins_num = df_bins_num.merge(rs, how = 'left', on = 'Variable')
df_bins_num = df_bins_num.sort_values(by=['Cluster', 'iv'], ascending = [True, False]).reset_index(drop = True)
df_bins_num['id'] = rs.groupby('Cluster').cumcount()+1

df_bins_num

IV > 0.15

- 0: LOAN, VALUE
- 1: DEBTINC, NINQ
- 2: DELINQ, DEROG
- 3: CLAGE
