==> notebooks/project/04.1_LENG_ModelingTesting.ipynb <==
import pandas as pd
import numpy as np
from joblib import load
import fs

MODELS_PROJECT_DIR = fs.open_fs("../../models/project")
PIPELINE_MODEL_DIR = MODELS_PROJECT_DIR.getsyspath("pipeline_model_tuning.joblib")

Save model

# load the model from disk
loaded_model = load(open(PIPELINE_MODEL_DIR, 'rb'))

# Diccionario con los datos
data = {'LOAN': 30.77702702702703,
        'MORTDUE': 44.6263399809302,
        'VALUE': 15.51946022509491,
        'REASON': "HomeImp",
        'JOB': "ProfExe",
        'YOJ': 10.0,
        'DEROG': 28.0,
        'DELINQ': 34.0,
        'CLAGE': 22.757786798573047,
        'NINQ': 10.0,
        'CLNO': 64.50704225352112,
        'DEBTINC': 26.638772794724}

# Convertir el diccionario a DataFrame, asumiendo que es una única fila
input_data = pd.DataFrame([data])

# Hacer una predicción con el modelo
prediction = loaded_model.predict(input_data)

# Mostrar el resultado de la predicción
print(f"Predicted class: {prediction[0]}")
